improvements to stumptown query and trend ui pages; fix issue with empty request for activity payloads

+++ code/com/jivesoftware/os/miru/miru-manage-deployable/src/main/java/com/jivesoftware/os/miru/manage/deployable/analytics/PaintWaveform.java
@@ -51,7 +51,7 @@ public class PaintWaveform {
         if (hits.length == 1) {
             hits = new double[] { hits[0], hits[0] };
         }
-        float ws = _w / (hits.length - 1);
+        float ws = (float) _w / (float) (hits.length - 1);
 
         g.setColor(color);
         for (int i = 1; i < hits.length; i++) {
